requirements.txt
python-dotenv
cryptography
aiosmtplib
portalocker
aiofiles
async_timeout
types-aiofiles
watchdog
pytest-asyncio
aiosqlite
mypy